pyproject.toml
[project]
name = "kwarapulseai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.24.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
